+++ RandomSkunk.Results/Operations/AsDisposable.cs
@@ -60,7 +60,7 @@ public static partial class ResultExtensions
 
         public AsyncDisposableResult(Result<TAsyncDisposable> source) => _source = source;
 
-        public async ValueTask DisposeAsync() => await _source.OnSuccessAsync(async value => await value.DisposeAsync().ConfigureAwait(false)).ConfigureAwait(false);
+        public async ValueTask DisposeAsync() => await _source.OnSuccess(async value => await value.DisposeAsync().ConfigureAwait(false)).ConfigureAwait(false);
     }
 
     private class AsyncDisposableMaybe<TAsyncDisposable> : IAsyncDisposable
@@ -70,6 +70,6 @@ public static partial class ResultExtensions
 
         public AsyncDisposableMaybe(Maybe<TAsyncDisposable> source) => _source = source;
 
-        public async ValueTask DisposeAsync() => await _source.OnSuccessAsync(async value => await value.DisposeAsync().ConfigureAwait(false)).ConfigureAwait(false);
+        public async ValueTask DisposeAsync() => await _source.OnSuccess(async value => await value.DisposeAsync().ConfigureAwait(false)).ConfigureAwait(false);
     }
 }